pyproject.toml
[project]
name = "rainforest-ecom"
version = "0.1.0"
description = ""
authors = [
    {name = "т"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "django-filter (>=25.1,<26.0)",
    "pre-commit (>=4.1.0,<5.0.0)",
    "django-request-logging (>=0.7.5,<0.8.0)",
    "django-ratelimit (>=4.1.0,<5.0.0)",
    "celery (>=5.4.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.10.0"
coverage = "^7.6.12"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "rainforest_ecom.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[tool.coverage.run]
branch = true
source = [
    "orders",
    "products",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "def __str__",
    "def __repr__",
]
